fix: update frontend PC components test route to /components
