+++ backend/pcmod/src/test/java/es/codeurjcstudents/pcmod/e2e/ui/ComponentsSystemTests.java
@@ -47,7 +47,7 @@ public class ComponentsSystemTests {
   @Test
   public void loadComponentsTest() throws InterruptedException {
 
-    driver.get("http://localhost:5173/");
+    driver.get("http://localhost:5173/components");
 
     wait.until(ExpectedConditions.presenceOfElementLocated(By.id("name-1")));
     String componentName = driver.findElement(By.id("name-1")).getText();
